add geo_functions module and move get_location,where_is_point from main to it
